+++ src/core.ts
